+++ plugin/src/main/kotlin/com/jtmnetwork/economy/entrypoint/api/impl/TransactionAPIImpl.kt
@@ -1,9 +1,10 @@
-package com.jtmnetwork.economy.entrypoint.api
+package com.jtmnetwork.economy.entrypoint.api.impl
 
 import com.google.inject.Inject
 import com.jtmnetwork.economy.core.domain.entity.Currency
 import com.jtmnetwork.economy.core.domain.entity.Transaction
 import com.jtmnetwork.economy.data.service.TransactionService
+import com.jtmnetwork.economy.entrypoint.api.TransactionAPI
 import org.bukkit.OfflinePlayer
 import org.bukkit.command.CommandSender
 import org.bukkit.entity.Player
