+++ plugin/src/main/kotlin/com/jtmnetwork/economy/entrypoint/module/WalletModule.kt
@@ -4,8 +4,6 @@ import com.google.inject.AbstractModule
 import com.jtmnetwork.economy.data.cache.WalletCache
 import com.jtmnetwork.economy.data.service.TransactionService
 import com.jtmnetwork.economy.data.service.WalletService
-import com.jtmnetwork.economy.entrypoint.api.WalletAPI
-import com.jtmnetwork.economy.entrypoint.api.WalletAPIImpl
 import com.jtmnetwork.economy.entrypoint.commands.PayCommands
 import com.jtmnetwork.economy.entrypoint.commands.RollbackCommands
 import com.jtmnetwork.economy.entrypoint.commands.WalletCommands
